task 6

+++ 0x00-hello_world/6-size.c
@@ -7,10 +7,10 @@
  */
 int main(void)
 {
-	printf("Size of a char: %zu byte(s)\n", sizeof(char));
-	printf("Size of an int: %zu bytes(s)\n", sizeof(int));
-	printf("Size of a long int: %zu bytes(s)\n", sizeof(long int));
-	printf("Size of a long long int: %zu bytes(s)\n", sizeof(long long int));
-	printf("Size of a float: %zu bytes(s)\n", sizeof(float));
+	printf("Size of a char: %zu byte(s)\n", (unsigned long)sizeof(char));
+	printf("Size of an int: %zu bytes(s)\n", (unsigned long)sizeof(int));
+	printf("Size of a long int: %zu bytes(s)\n", (unsigned long)sizeof(long int));
+	printf("Size of a long long int: %zu bytes(s)\n", (unsigned long)sizeof(long long int));
+	printf("Size of a float: %zu bytes(s)\n", (unsigned long)sizeof(float));
 	return (0);
 }
